+++ libraries/PCA9553/PCA9553.cpp
@@ -2,7 +2,7 @@
 //    FILE: PCA9553.cpp
 //  AUTHOR: Rob Tillaart
 //    DATE: 2023-07-16
-// VERSION: 0.1.0
+// VERSION: 0.2.0
 // PURPOSE: Arduino library for for I2C PCA9553 4 channel PWM
 //     URL: https://github.com/RobTillaart/PCA9553
 
@@ -18,6 +18,8 @@ PCA9553::PCA9553(const uint8_t deviceAddress, TwoWire *wire)
 {
   _address = deviceAddress;
   _wire    = wire;
+  _outputCount = 4;
+  _error = PCA9553_OK;
 }
 
 
@@ -53,9 +55,30 @@ bool PCA9553::isConnected()
 }
 
 
-uint8_t PCA9553::channelCount()
+uint8_t PCA9553::reset()
 {
-  return 4;
+  //  not most efficient
+  setPrescaler(0, 0);   //  44 Hz
+  setPrescaler(1, 0);   //  44 Hz
+  setPWM(0, 128);       //  50%
+  setPWM(1, 128);       //  50%
+  for (int pin = 0; pin < _outputCount; pin++)
+  {
+    setOutputMode(pin, 0);  //  LOW
+  }
+  return PCA9553_OK;
+}
+
+
+uint8_t PCA9553::getAddress()
+{
+  return _address;
+}
+
+
+uint8_t PCA9553::outputCount()
+{
+  return _outputCount;
 }
 
 
@@ -69,31 +92,42 @@ uint8_t PCA9553::getInput()
 }
 
 
+void PCA9553::pinMode(uint8_t pin, uint8_t mode)
+{
+  if (mode != OUTPUT) setOutputMode(pin, PCA9553_MODE_HIGH);
+}
+
+
+void PCA9553::digitalWrite(uint8_t pin, uint8_t val)
+{
+  if (val == LOW) setOutputMode(pin, PCA9553_MODE_LOW);
+  else            setOutputMode(pin, PCA9553_MODE_HIGH);
+}
+
+
+uint8_t PCA9553::digitalRead(uint8_t pin)
+{
+  uint8_t value = readReg(PCA9553_INPUT);
+  if ((value >> pin) & 0x01) return HIGH;
+  return LOW;
+}
+
+
 /////////////////////////////////////////////////////
 //
 //  PRESCALERS
 //
-void PCA9553::setPrescaler0(uint8_t psc)
+void PCA9553::setPrescaler(uint8_t gen, uint8_t psc)
 {
-  writeReg(PCA9553_PSC0, psc);
+  if (gen == 0) writeReg(PCA9553_PSC0, psc);
+  else          writeReg(PCA9553_PSC1, psc);
 }
 
 
-uint8_t PCA9553::getPrescaler0()
+uint8_t PCA9553::getPrescaler(uint8_t gen)
 {
-  return readReg(PCA9553_PSC0);
-}
-
-
-void PCA9553::setPrescaler1(uint8_t psc)
-{
-  writeReg(PCA9553_PSC1, psc);
-}
-
-
-uint8_t PCA9553::getPrescaler1()
-{
-  return readReg(PCA9553_PSC1);
+  if (gen == 0) return readReg(PCA9553_PSC0);
+  else          return readReg(PCA9553_PSC1);
 }
 
 
@@ -101,57 +135,68 @@ uint8_t PCA9553::getPrescaler1()
 //
 //  PWM
 //
-void PCA9553::setPWM0(uint8_t pwm)
+void PCA9553::setPWM(uint8_t gen, uint8_t pwm)
 {
-  writeReg(PCA9553_PWM0, pwm);
+  if (gen == 0) writeReg(PCA9553_PWM0, pwm);
+  else          writeReg(PCA9553_PWM1, pwm);
 }
 
 
-uint8_t PCA9553::getPWM0()
+uint8_t PCA9553::getPWM(uint8_t gen)
 {
-  return readReg(PCA9553_PWM0);
-}
-
-
-void PCA9553::setPWM1(uint8_t pwm)
-{
-  writeReg(PCA9553_PWM1, pwm);
-}
-
-
-uint8_t PCA9553::getPWM1()
-{
-  return readReg(PCA9553_PWM1);
+  if (gen == 0) return readReg(PCA9553_PWM0);
+  else          return readReg(PCA9553_PWM1);
 }
 
 
 /////////////////////////////////////////////////////
 //
-//  LED SOURCE SELECTOR
+//  OUTPUT MODE - MUX SELECTION
 //
-bool PCA9553::setLEDSource(uint8_t led, uint8_t source)
+uint8_t PCA9553::setOutputMode(uint8_t pin, uint8_t mode)
 {
-  if (led > 3) return false;
-  if (source > 3) return false;
+  if (pin >= _outputCount)
+  {
+    _error = PCA9553_ERR_CHAN;
+    return _error;
+  }
+  if (mode > 3)
+  {
+    _error = PCA9553_ERR_MODE;
+    return _error;
+  }
 
-  uint8_t val = source << (led * 2);
-  uint8_t ledSelect = readReg(PCA9553_LS0);
-  ledSelect &= ~(0x03 << (led * 2));
-  ledSelect |= val;
+  uint8_t reg = PCA9553_LS0;
+  uint8_t ledSelect = readReg(reg);
+  ledSelect &= ~(0x03 << (pin * 2));
+  ledSelect |= (mode << (pin * 2));
 
-  writeReg(PCA9553_LS0, ledSelect);
-  return true;
+  return writeReg(reg, ledSelect);
 }
 
 
-uint8_t PCA9553::getLEDSource(uint8_t led)
+uint8_t PCA9553::getOutputMode(uint8_t pin)
 {
-  if (led > 3) return PCA9553_ERROR;
+  if (pin >= _outputCount)
+  {
+    _error = PCA9553_ERR_CHAN;
+    return _error;
+  }
 
-  uint8_t ledSelect = readReg(PCA9553_LS0);
-  uint8_t source = (ledSelect >> (led * 2)) & 0x03;
-  return source;
+  uint8_t reg = PCA9553_LS0;
+  uint8_t ledSelect = readReg(reg);
+  uint8_t mode = (ledSelect >> (pin * 2)) & 0x03;
+  return mode;
 }
+
+
+uint8_t PCA9553::getLastError()
+{
+  uint8_t e = _error;
+  _error = PCA9553_OK;
+  return e;
+}
+
 
 
 /////////////////////////////////////////////////////
@@ -166,7 +211,7 @@ uint8_t PCA9553::writeReg(uint8_t reg, uint8_t value)
   _error = _wire->endTransmission();
 
   if (_error == 0) _error = PCA9553_OK;
-  else _error = PCA9553_ERROR;
+  else _error = PCA9553_ERR_I2C;
   return _error;
 }
 
@@ -176,10 +221,15 @@ uint8_t PCA9553::readReg(uint8_t reg)
   _wire->beginTransmission(_address);
   _wire->write(reg);
   _error = _wire->endTransmission();
-
+  if (_error == 0) _error = PCA9553_OK;
+  else
+  {
+    _error = PCA9553_ERR_I2C;
+    return 0;
+  }
   if (_wire->requestFrom(_address, (uint8_t)1) != 1)
   {
-    _error = PCA9553_ERROR;
+    _error = PCA9553_ERR_I2C;
     return 0;
   }
   _error = PCA9553_OK;