+++ libraries/PCA9553/PCA9553.h
@@ -3,7 +3,7 @@
 //    FILE: PCA9553.h
 //  AUTHOR: Rob Tillaart
 //    DATE: 2023-07-16
-// VERSION: 0.1.0
+// VERSION: 0.2.0
 // PUPROSE: Arduino library for for I2C PCA9553 4 channel PWM
 //     URL: https://github.com/RobTillaart/PCA9553
 
@@ -12,7 +12,7 @@
 #include "Wire.h"
 
 
-#define PCA9553_LIB_VERSION         (F("0.1.0"))
+#define PCA9553_LIB_VERSION         (F("0.2.0"))
 
 
 //  REGISTERS
@@ -24,11 +24,14 @@
 #define PCA9553_LS0         0x05
 
 
-//  LED SOURCE MASKS
-//  TODO
+//  MUX OUTPUT MODES
+#define PCA9553_MODE_LOW      0
+#define PCA9553_MODE_HIGH     1
+#define PCA9553_MODE_PWM0     2
+#define PCA9553_MODE_PWM1     3
 
 
-//  ERROR CODES
+//  ERROR CODES (not all used yet)
 #define PCA9553_OK          0x00
 #define PCA9553_ERROR       0xFF
 #define PCA9553_ERR_WRITE   0xFE
@@ -52,39 +55,38 @@ public:
 #endif
   bool     begin();
   bool     isConnected();
+  uint8_t  reset();
 
-  uint8_t  channelCount();
+  uint8_t  getAddress();
+  uint8_t  outputCount();
 
-  //  input register, only lower 4 bits
+  //  GPIO
   uint8_t  getInput();
-
+  void     pinMode(uint8_t pin, uint8_t mode);
+  void     digitalWrite(uint8_t pin, uint8_t value);
+  uint8_t  digitalRead(uint8_t pin);
 
   //  PRESCALERS
-  void     setPrescaler0(uint8_t psc = 255);
-  uint8_t  getPrescaler0();
-
-  void     setPrescaler1(uint8_t psc = 255);
-  uint8_t  getPrescaler1();
-
+  void     setPrescaler(uint8_t generator, uint8_t prescaler = 0);
+  uint8_t  getPrescaler(uint8_t generator);
 
   //  PWM
-  void     setPWM0(uint8_t pwm = 128);
-  uint8_t  getPWM0();
-
-  void     setPWM1(uint8_t pwm = 128);
-  uint8_t  getPWM1();
+  void     setPWM(uint8_t generator, uint8_t pwm = 128);
+  uint8_t  getPWM(uint8_t generator);
 
 
-  //  LED SOURCE SELECTOR
-  //  led = 0..3
+  //  OUTPUT MODE - MUX SELECTION
+  //  pin:   0..3
+  //  mode:
+  //    0 = output is set LOW (LED on)
+  //    1 = output is set high-impedance (LED off; default)
+  //    2 = output blinks at PWM0 rate
+  //    3 = output blinks at PWM1 rate
+  uint8_t  setOutputMode(uint8_t pin, uint8_t mode);
+  uint8_t  getOutputMode(uint8_t pin);
 
-  //  source:
-  //    00 = output is set LOW (LED on)
-  //    01 = output is set high-impedance (LED off; default)
-  //    10 = output blinks at PWM0 rate
-  //    11 = output blinks at PWM1 rate
-  bool     setLEDSource(uint8_t led, uint8_t source);
-  uint8_t  getLEDSource(uint8_t led);
+  //  ERROR
+  uint8_t  getLastError();
 
 
 private:
@@ -92,7 +94,9 @@ private:
   uint8_t  readReg(uint8_t reg);
 
   uint8_t  _address;
+  uint8_t  _outputCount;
   uint8_t  _error;
+
   TwoWire* _wire;
 };